+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.4.0 - 2026-03-01
+### Updated
+- Adapted to canonical `address.display` field.
+- Removed dependency on `address.line1`.
+
 ## 0.3.0 - 2026-02-26
 ### Added
 - JSON-LD ItemList structured data for listing carousel.